RAG_AI/requirements.txt
# Core ML dependencies
torch>=2.1.0
transformers>=4.35.2
sentence-transformers>=2.2.2
huggingface-hub>=0.19.0

# Database dependencies
pymongo>=4.6.0
qdrant-client>=1.6.4

# Web and API
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
gradio>=4.7.1

# Data processing
beautifulsoup4>=4.12.2
requests>=2.31.0
pandas>=2.1.3
numpy>=1.26.2

# Development tools
pytest>=7.4.3
black>=23.11.0
peft>=0.5.0
bitsandbytes>=0.41.1

# Web scraping
selenium>=4.15.2
webdriver-manager>=4.0.1

# ROS2 specific
PyGithub>=2.1.1
markdown>=3.5.1
accelerate>=0.26.0

# Ollama client
ollama>=0.1.6

# ClearML
clearml>=1.14.1

# Videos
youtube-transcript-api>=0.6.1
pytube>=15.0.0

#---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
# Incomplete requirements

# # Core ML dependencies
# --find-links https://download.pytorch.org/whl/torch_stable.html
# torch==2.1.0+cpu
# transformers==4.35.2
# sentence-transformers==2.2.2
# huggingface-hub==0.16.4  # Added specific version for compatibility

# # Database dependencies
# pymongo==4.6.0
# qdrant-client==1.6.4

# # Web and API
# fastapi==0.104.1
# uvicorn==0.24.0
# python-dotenv==1.0.0
# gradio==4.7.1

# # Data processing
# beautifulsoup4==4.12.2
# requests==2.31.0
# pandas==2.1.3
# numpy==1.26.2

# # Development tools
# jupyter==1.0.0
# notebook==7.0.6
# pytest==7.4.3
# black==23.11.0

# # Web scraping
# selenium==4.15.2
# webdriver_manager==4.0.1

# # Video processing
# youtube_transcript_api==0.6.1
# pytube==15.0.0

# # ML Experiment tracking
# clearml==1.16.5

# # Make sure pip is updated
# pip>=23.0.1